+++ src/main/java/org/stepup/cinesquareapis/report/model/UserScoreRequest.java
@@ -3,11 +3,9 @@ package org.stepup.cinesquareapis.report.model;
 import lombok.Data;
 import org.stepup.cinesquareapis.report.entity.UserScore;
 
-import java.time.LocalDateTime;
-
 @Data
 public class UserScoreRequest {
-    private Double score;
+    private float score;
 
     public UserScore toEntity(int userId, int movieId) {
         return new UserScore().builder()
